requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cu111
torchvision==0.9.0+cu111
opencv-python==4.12.0.88
numpy==1.24.4
scipy==1.10.1
lpips==0.1.4
imageio==2.35.1
pillow==10.4.0
matplotlib==3.7.5
scikit-learn==1.3.2
pandas==2.0.3
